Revert "Deploy dita-ot/docs@bdd28ba to 'dev' docs"
The `clean.rb` script in the docs repo is supposed to reset files where only the commit metadata has changed, but this was disabled in dita-ot/docs@8ee3bf11 to prevent side-effects when commit metadata generation stopped working with the 2.3 release, so the historic commit IDs for each topic were replaced with the latest docs commit ID in 9873136.

This reverts commit 9873136fb548c30066fac40132737995827cc4b0.

The `clean.rb` script was re-enabled in dita-ot/docs@0e5c4d67, so subsequent deployments should be cleaned as intended.

Signed-off-by: Roger Sheen <roger@infotexture.net>