Add splash layout & move navbar margin to #content
The splash layout is a copy of the base layout that wraps the generated page {{ content }} in an HTML5 <main> element instead of the top-level container div & nested row that the base layout applies.

This enables fluid jumbotron components to extend to the full width of the viewport.

For a seamless appearance, the margin between the navbar and the main page content is also moved from the bottom of the navbar to the top of the base layout container.

Signed-off-by: Roger Sheen <roger@infotexture.net>