Collapse navbar menu at smaller breakpoints
Below the large breakpoint (992px width), the Download, Plug-ins & DITA-OT Day links are hidden, but the other menu items remain visible for easy access on medium-sized devices (tablets, 768px and up).

On smaller devices like phones, the collapse plugin is used to hide the links and show a menu button to toggle the collapsed content.

Signed-off-by: Roger Sheen <roger@infotexture.net>