Replace reference-style links with inline links
The `DCO.md` file is also included in the `contributing.md` file, and both use Markdown reference-style links, but they each have different link targets for references 1–3.

In the context of the including file, the links in the included file inherit the parent file’s link targets and point to the wrong URLs.

Using inline links in the included file ensures that those links point to the correct URLs regardless of the context.

Signed-off-by: Roger Sheen <roger@infotexture.net>